Rename `logical_meter` package to `_logical_meter` to make it private
... using a patch helpfully provided by Luca to fix the cyclic
dependencies arising from the renaming:
https://github.com/frequenz-floss/frequenz-sdk-python/pull/66#discussion_r1032450423

The only change that went into the patch was an isort fix to reorder
imports.
