+++ src/frequenz/sdk/timeseries/__init__.py
@@ -10,11 +10,13 @@ A timeseries is a stream (normally an async iterator) of
 This module provides tools to operate on timeseries.
 """
 
+from ._logical_meter import LogicalMeter
 from ._resampler import GroupResampler, Resampler, ResamplingFunction
 from ._sample import Sample
 
 __all__ = [
     "GroupResampler",
+    "LogicalMeter",
     "Resampler",
     "ResamplingFunction",
     "Sample",
