+++ src/frequenz/sdk/timeseries/_logical_meter/_formula_builder.py
@@ -6,7 +6,8 @@
 
 from frequenz.channels import Receiver, Sender
 
-from ...actor import ChannelRegistry, ComponentMetricRequest
+from ...actor._channel_registry import ChannelRegistry
+from ...actor._data_sourcing.microgrid_api_source import ComponentMetricRequest
 from ...microgrid.component import ComponentMetricId
 from .._sample import Sample
 from ._formula_engine import FormulaEngine
