+++ src/frequenz/sdk/timeseries/_logical_meter/_logical_meter.py
@@ -9,7 +9,8 @@ from typing import Dict, List
 
 from frequenz.channels import Broadcast, Receiver, Sender
 
-from ...actor import ChannelRegistry, ComponentMetricRequest
+from ...actor._channel_registry import ChannelRegistry
+from ...actor._data_sourcing.microgrid_api_source import ComponentMetricRequest
 from ...microgrid.component import ComponentMetricId
 from .._sample import Sample
 from ._formula_builder import FormulaBuilder
